+++ openfisca_us/variables/irs/income/deductions/deductions.py
@@ -1,0 +1,1 @@
+from openfisca_us.model_api import *
