+++ openfisca_us/variables/irs/outputs.py
@@ -23,28 +23,6 @@ class filer_sey(Variable):
         return tax_unit_non_dep_sum("sey", tax_unit, period)
 
 
-class niit(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Net Investment Income Tax"
-    unit = USD
-    documentation = "Net Investment Income Tax from Form 8960"
-
-    def formula(tax_unit, period, parameters):
-        NII_ELEMENTS = [
-            "filer_e00300",
-            "filer_e00600",
-            "c01000",
-            "filer_e02000",
-        ]
-        nii = max_(0, add(tax_unit, period, NII_ELEMENTS))
-        niit = parameters(period).irs.investment.net_inv_inc_tax
-        threshold = niit.threshold[tax_unit("mars", period)]
-        base = min_(nii, max_(0, tax_unit("c00100", period) - threshold))
-        return niit.rate * base
-
-
 class combined(Variable):
     value_type = float
     entity = TaxUnit
@@ -192,23 +170,6 @@ class expanded_income(Variable):
         )
 
 
-class iitax(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    unit = USD
-    label = "Federal income tax"
-    documentation = "Total federal individual income tax liability."
-
-    def formula(tax_unit, period, parameters):
-        return tax_unit(
-            "income_tax_before_refundable_credits", period
-        ) - tax_unit("income_tax_refundable_credits", period)
-
-
-income_tax = variable_alias("income_tax", iitax)
-
-
 class num(Variable):
     value_type = int
     entity = TaxUnit
@@ -223,19 +184,6 @@ class othertaxes(Variable):
     definition_period = YEAR
     documentation = "Other taxes: sum of niit, e09700, e09800 and e09900 (included in c09200)"
     unit = USD
-
-
-class income_tax_refundable_credits(Variable):
-    value_type = float
-    entity = TaxUnit
-    definition_period = YEAR
-    label = "Income tax refundable credits"
-    documentation = "Total refundable income tax credits"
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        credits = parameters(period).irs.credits.refundable
-        return add(tax_unit, period, credits)
 
 
 class sep(Variable):
@@ -255,81 +203,6 @@ class filer_sey(Variable):
 
     def formula(tax_unit, period, parameters):
         return tax_unit_non_dep_sum("sey", tax_unit, period)
-
-
-class basic_standard_deduction(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Basic standard deduction"
-    definition_period = YEAR
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        std = parameters(period).irs.deductions.standard
-        mars = tax_unit("mars", period)
-        midr = tax_unit("midr", period)
-
-        c15100_if_dsi = max_(
-            std.dependent.additional_earned_income
-            + tax_unit("filer_earned", period),
-            std.dependent.amount,
-        )
-        basic_if_dsi = min_(std.amount[mars], c15100_if_dsi)
-        basic_if_not_dsi = where(midr, 0, std.amount[mars])
-        return where(tax_unit("dsi", period), basic_if_dsi, basic_if_not_dsi)
-
-
-class aged_blind_extra_standard_deduction(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Aged and blind standard deduction"
-    definition_period = YEAR
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        std = parameters(period).irs.deductions.standard
-        mars = tax_unit("mars", period)
-        mars_type = mars.possible_values
-        blind_head = tax_unit("blind_head", period) * 1
-        blind_spouse = tax_unit("blind_spouse", period) * 1
-        aged_head = (
-            tax_unit("age_head", period) >= std.aged_or_blind.age_threshold
-        ) * 1
-        aged_spouse = (
-            (mars == mars_type.JOINT)
-            & (
-                tax_unit("age_spouse", period)
-                >= std.aged_or_blind.age_threshold
-            )
-        ) * 1
-        num_extra_stded = blind_head + blind_spouse + aged_head + aged_spouse
-        return num_extra_stded * std.aged_or_blind.amount[mars]
-
-
-class standard(Variable):
-    value_type = float
-    entity = TaxUnit
-    label = "Standard deduction (zero for itemizers)"
-    definition_period = YEAR
-    unit = USD
-
-    def formula(tax_unit, period, parameters):
-        # Calculate basic standard deduction
-        basic_stded = tax_unit("basic_standard_deduction", period)
-        charity = parameters(period).irs.deductions.itemized.charity
-        mars = tax_unit("mars", period)
-        midr = tax_unit("midr", period)
-        mars_type = mars.possible_values
-
-        # Calculate extra standard deduction for aged and blind
-        extra_stded = tax_unit("aged_blind_extra_standard_deduction", period)
-
-        # Calculate the total standard deduction
-        standard = basic_stded + extra_stded
-        standard = where((mars == mars_type.SEPARATE) & midr, 0, standard)
-        return standard + charity.allow_nonitemizers * min_(
-            tax_unit("c19700", period), charity.nonitemizers_max
-        )
 
 
 class surtax(Variable):
@@ -1531,7 +1404,6 @@ class pre_c04600(Variable):
             0,
             tax_unit("xtot", period) * exemption.amount,
         )
-
 
 
 class filer_setax(Variable):
