+++ openfisca_us/variables/irs/tax/payroll.py
@@ -31,7 +31,6 @@ class employee_payrolltax(Variable):
         return tax_unit("payrolltax", period) * 0.5
 
 
-
 class ptax_amc(Variable):
     value_type = float
     entity = TaxUnit
@@ -270,4 +269,25 @@ class extra_payrolltax(Variable):
             extra_ss_income
             * not_(tax_unit.members("is_tax_unit_dependent", period))
             * ss.rate
+        )
+
+
+class social_security_taxes(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "Social security taxes"
+    unit = "currency-USD"
+    documentation = "Total employee-side social security taxes"
+    definition_period = YEAR
+
+    def formula(tax_unit, period, parameters):
+        employee_payroll_tax = 0.5 * tax_unit("ptax_was", period)
+        self_employed_tax = tax_unit("c03260", period)
+        unreported_payroll_tax = aggr(tax_unit, period, ["e09800"])
+        excess_payroll_tax_withheld = aggr(tax_unit, period, ["e11200"])
+        return (
+            employee_payroll_tax
+            + self_employed_tax
+            + unreported_payroll_tax
+            - excess_payroll_tax_withheld
         )
